refactor(Navbar): move Navbar to separate file
